microservice-client: create view /info-Patient  by id and method with  @GetMapping('/info-patient/{id}')

+++ microservice-client/src/main/java/com/mclient/microserviceclient/proxy/IMicroservicePatientsProxy.java
@@ -21,10 +21,10 @@ public interface IMicroservicePatientsProxy {
 	PatientBean createPatient(@Valid @RequestBody PatientBean patient);
 
 	@PutMapping("/MICROSERVICE-PATIENTS/patient/modification")
-	PatientBean updateOnePatientById(@Valid @RequestBody PatientBean patient, @RequestParam long id);
+	PatientBean updateOnePatientById(@Valid @RequestBody PatientBean patient, @RequestParam Integer id);
 
 	@GetMapping("/MICROSERVICE-PATIENTS/patient/info-patient/{id}")
-	PatientBean getPatientById(@PathVariable long id) ;
+	PatientBean getPatientById(@PathVariable Integer  id) ;
 
 	@GetMapping("MICROSERVICE-PATIENTS/patient/list")
 	List<PatientBean> getAllPatients() ;
